mp3: rework Xing/VBRI header parsing to be more strict; parse LAME Xing headers; extract the bitrate info from VBRI headers. (Fixes issue # 182)
Instead of scanning the file from the guessed offset and searching
for the header identifier only look at the correct location based
on the mpeg header location and content. This reduces the risk that
we interpret random data as a xing header.

While at it also read lame Xing headers which start with "Info" but are
compatible besides that and extract the real bitrate info from VBRI
headers.